+++ launcher/resource.h
@@ -56,6 +56,7 @@
 #define IDC_SWAP_GRENADE_CONTROLS       1047
 #define IDC_PLAYER_JOIN_BEEP_CHECK      1048
 #define IDC_MESH_STATIC_LIGHTING_CHECK  1049
+#define IDC_AUTOSAVE_CHECK              1050
 
 // Next default values for new objects
 //
@@ -63,7 +64,7 @@
 #ifndef APSTUDIO_READONLY_SYMBOLS
 #define _APS_NEXT_RESOURCE_VALUE        105
 #define _APS_NEXT_COMMAND_VALUE         32771
-#define _APS_NEXT_CONTROL_VALUE         1050
+#define _APS_NEXT_CONTROL_VALUE         1051
 #define _APS_NEXT_SYMED_VALUE           101
 #endif
 #endif